[move-2] Making enum types ready for prod
This PR does a number of things to get enum types ready for production. Those are more or less related via e2e tests:

- Implement missing _compatibility check_ for enums. Since the old compatibility checker was based on deprecated code, created a new one directly working on file format and views. This is feature gated: the old checker is in `compatibility_legacy.rs` and will be invoked unless feature flag for new one is set.
- Makes bytecode _v7_ the default bytecode version for the VM. The actual version on nodes is determined by on-chain config. Removed the VERSION_NEXT boilerplate for simplification.
- Added explicit feature gating via on-chain config for (a) enum types (b) access control.
- Extend CompilerOptions to support experiments passed on to the v2 compiler; this allows to use access control from e2e move tests.

The new compatibility checker is tested via existing e2e tests in `code_publishing.rs` and a new test specifically for enums in `enum_upgrade.rs`. (While the new checker is off on-chain, it is on in e2e tests.) The feature gating is tested in the new test `move_feature_gating.rs`. All e2e tests now operate with v7 bytecode, so will nodes do (but new features in v7 are turned off on-chain).